Add 'run_ctest' script and update commands in recipe.yaml
This commit adds a 'run_ctest' shell script to provide an easily accessible test runner for Code_Aster. The 'run_aster' script is also added for complete functionality. The 'asrun' commands used for testing in 'recipe.yaml' have been replaced by the newly added scripts for ease of running tests. Note that this also requires the addition of 'asrun' to the packages required for building. The 'frontend' dependency has been commented out due to these changes.

In the 'build.sh' file, the pip installation of 'asrun' was commented out and the copying of the new scripts into the build directory was added.

Finally, 'krande' and 'conda-forge' channels were added to the GitHub workflows CI configuration for 'code_aster'. These channels were required for the new dependencies introduced in this commit.