Refactor source code to improve initialization and execution process
The changes include refactoring of the initialization and execution process of code_aster to throw a more descriptive error when the import fails to improve the developer experience. Changes in the test_import.py file allow more granular testing and increased modularity. The recipe.yaml, code_aster_activate.sh, build.sh, and wafcfg_conda.py files have been modified to accommodate the changes in the initialization process and improve the build process. The mfront's build.sh file was also modified to improve the build process. In the file ci-code-aster.yml, the testing process was upgraded by enabling tests for Python version 3.10 on Linux. In recipe.yaml of mfront, pin_subpackage was added to ensure compatibility with the environment. Furthermore, the asrun interface for code_aster services has been added.