Update CI workflow to upload build artifacts to anaconda
Changes have been made to the GitHub Actions workflow ci-code-aster.yml. The changes include adding a step for uploading built package artifacts to anaconda cloud for windows and linux platforms. The changes have been incorporated in order to ease the accessibility of built packages for the users. It also includes the creation of a separate directory to store the build artifacts which previously was not being used. This will streamline the process of accessing and managing these artifacts.