Refactor workflow file to use environment variable for paths
The ci-code-aster.yml workflow file has been refactored to replace all instances of `github.workspace` with a new environment variable `WORK_DIR`. The change was necessary to increase readability and to make maintenance easier, especially when paths need to be updated in the future. This approach makes our workflow more flexible as we can now alter the working directory by simply modifying a single environment variable.