Add support for PETSc library and adjust paths in activation scripts
This commit includes the addition of environment variables for the PETSc library in the build.sh script, which means that the code_aster software is now able to use this library for certain computations. Also, the prefix for various environment variables in the code_aster_activate.sh and code_aster_deactivate.sh scripts has been changed from $PREFIX to $CONDA_PREFIX. This aligns the scripts with convention for conda environment activation scripts and should help prevent issues with variables being set or unset incorrectly when the environment is activated or deactivated.