Update build commands and remove unused dependencies
Updated the build commands in the ci-mfront.yml and ci-libmed.yml workflow files to use 'pkg-format=2' for better package management. This change necessitated the removal of the 'perl' and 'm4' dependencies in the mfront and libmed recipes, which are not used in the updated build process. Also reduced the exporting of unnecessary variables in the libmed build.sh file and simplified the process instead to straight commands 'configure', 'make' and 'make install' for easier maintenance of the script.