Update artifact upload file extension in ci-libmed workflow
This change modifies the 'upload to anaconda' step in the ci-libmed Github workflow. The script previously was searching for  '.conda' files to upload. However, the generated artifact files are in '.tar.bz2' format, therefore this step has been changed to reflect the correct file extension. This will ensure correct and successful uploads to anaconda.