Update upload process and build version
This commit is updating the process of uploading packages to Anaconda in the ci-code-aster-batch.yml file. Instead of iterating over different file extensions, it now checks directly if the package is of type 'boa' and adjusts the path accordingly. This makes the upload action more flexible and potentially faster.

The build number in the recipe.yaml in code_aster directory has also been incremented. The change in the build version indicates that there was an update to the package code.