Update numpy version to 1.23
The numpy package version in various conda_build_config.yaml files and in the modhelper.py script has been updated from 1.22 to 1.23. This was necessary to ensure compatibility with recent dependencies and to maintain updated and synchronized package versions. This change was incorporated across all projects relying on the numpy package within this repository. The modhelper.py script was also adjusted to facilitate future updates to package versions in a more generalized approach.