Update build string format and bump version
The build string format for the 'ifcopenshell' and 'gmsh' packages in the Conda packaging were restructured to fix the inconsistent naming issue. The string now includes 'py' before 'CONDA_PY' for better readability and standardization across the packages. The version bump in the CI workflows for these packages has been reset to 0 for the upcoming new release.