Update library dependencies in build scripts
Updated petsc build configurations to include parmetis library and add numpy as a library dependency for parmetis. Also, made adjustments in ci-code-aster-batchv2 configuration.

The change in petsc build configuration was necessary because the newly added '-lparmetis' flag is required for petsc's compilation and operation. Addition of numpy as a host and run-time dependency in the parmetis's meta.yaml allows numpy-based operations in parmetis package. The ci-code-aster-batchv2 configuration file has been updated to decrease the version bump count from 1 to 0 for proper version control. Lastly, numpy version is explicitly set to 1.23 in parmetis build configuration to avoid version conflicts.